requirements.txt
mysql-connector-python==8.0.18
Pillow==6.2.1
protobuf==3.10.0
six==1.12.0
